Refactor to class
